Refactor: 물섭취 카드 페이지 분리
